feat(User): Repositories of UserRepository and RepositoryRepository created

+++ src/main/java/com/devspotlight/devspotlight/repository/UserRepository.java
@@ -3,6 +3,5 @@ package com.devspotlight.devspotlight.repository;
 import com.devspotlight.devspotlight.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
 }
